fix: Fix nav item rename error
